requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
faiss-cpu
pytesseract
pdf2image
Pillow
torch
transformers
sentence-transformers
numpy
streamlit
fastapi
uvicorn
pydantic
unicodedata2
regex
python-dotenv
requests
typing-extensions